feat(zudo-doc): ship pure-helper content components to the package (#2327)
Move the near-zero-coupling content helpers into @takazudo/zudo-doc as part
of the package-first migration (epic #2321, S4):

- makeAdmonition  -> @takazudo/zudo-doc/content-admonition
- CodeGroup       -> @takazudo/zudo-doc/code-group (glue over Tabs/TabItem)
- MathBlock       -> @takazudo/zudo-doc/math-block (katex-only; katex added as
                     an optional peerDependency)
- extractHeadings -> @takazudo/zudo-doc/extract-headings (the single settings
                     read is now passed in via opts.{tocMinDepth,tocMaxDepth,
                     strategy} — no project singleton import)
- Settings + config interfaces -> @takazudo/zudo-doc/settings (pure types)

The showcase now re-exports/imports these from the package; the project keeps
its concrete src/config/settings.ts doing `satisfies Settings` (data stays
project-side). The pages/lib/_extract-headings.ts host wrapper re-reads the
project settings and forwards them so existing call sites are unchanged.

New export subpaths appended to packages/zudo-doc/package.json (after the
existing JS subpath group, before the .css entries, per the S2 convention);
new src/** files are auto-globbed by tsup. Template counterparts intentionally
keep the full implementations for now — allowlisted in .template-drift-allowlist
pending a follow-up generator-sync wave.

build + typecheck (pnpm check) + package tests (596 passed) all green.
